+++ src/tweenable.rs
@@ -116,19 +116,25 @@ impl AnimClock {
         }
     }
 
-    fn record_completions(&mut self, times_completed: u32) {
-        self.times_completed = self.times_completed.saturating_add(times_completed);
-    }
-
-    fn tick(&mut self, tick: Duration) -> u32 {
+    fn tick(&mut self, tick: Duration) -> (TweenState, u32) {
         let duration = self.duration.as_nanos();
+        let prev_times_completed = self.elapsed.as_nanos() / duration;
 
-        let before = self.elapsed.as_nanos() / duration;
         self.elapsed = self.elapsed.saturating_add(tick);
-        if let TotalDuration::Finite(duration) = self.total_duration {
-            self.elapsed = self.elapsed.min(duration);
-        }
-        (self.elapsed.as_nanos() / duration - before) as u32
+        let state = if let TotalDuration::Finite(duration) = self.total_duration {
+            if self.elapsed >= duration {
+                self.elapsed = duration;
+                TweenState::Completed
+            } else {
+                TweenState::Active
+            }
+        } else {
+            TweenState::Active
+        };
+
+        let times_completed = (self.elapsed.as_nanos() / duration - prev_times_completed) as u32;
+        self.times_completed = self.times_completed.saturating_add(times_completed);
+        (state, times_completed)
     }
 
     fn set_progress(&mut self, progress: f32) {
@@ -136,7 +142,12 @@ impl AnimClock {
     }
 
     fn progress(&self) -> f32 {
-        self.elapsed.as_secs_f32() / self.duration.as_secs_f32()
+        if let TotalDuration::Finite(total_duration) = self.total_duration {
+            if self.elapsed >= total_duration {
+                return 1.;
+            }
+        }
+        (self.elapsed.as_secs_f32() / self.duration.as_secs_f32()).fract()
     }
 
     fn state(&self) -> TweenState {
@@ -493,12 +504,16 @@ impl<T> Tweenable<T> for Tween<T> {
         }
 
         // Tick the animation clock
-        let times_completed = self.clock.tick(delta);
-        self.clock.record_completions(times_completed);
-        if self.clock.strategy == RepeatStrategy::MirroredRepeat && times_completed & 1 != 0 {
+        let (state, times_completed) = self.clock.tick(delta);
+        let (progress, times_completed_for_direction) = match state {
+            TweenState::Active => (self.progress(), times_completed),
+            TweenState::Completed => (1., times_completed.max(1) - 1), // ignore last
+        };
+        if self.clock.strategy == RepeatStrategy::MirroredRepeat
+            && times_completed_for_direction & 1 != 0
+        {
             self.direction = !self.direction;
         }
-        let progress = self.progress();
 
         // Apply the lens, even if the animation finished, to ensure the state is
         // consistent
@@ -522,7 +537,7 @@ impl<T> Tweenable<T> for Tween<T> {
             }
         }
 
-        self.clock.state()
+        state
     }
 
     fn times_completed(&self) -> u32 {
@@ -530,6 +545,20 @@ impl<T> Tweenable<T> for Tween<T> {
     }
 
     fn rewind(&mut self) {
+        if self.clock.strategy == RepeatStrategy::MirroredRepeat {
+            // In mirrored mode, direction alternates each loop. To reset to the original
+            // direction on Tween creation, we count the number of completions, ignoring the
+            // last one if the Tween is currently in TweenState::Completed because that one
+            // freezes all parameters.
+            let mut times_completed = self.clock.times_completed;
+            if self.clock.state() == TweenState::Completed {
+                debug_assert!(times_completed > 0);
+                times_completed -= 1;
+            }
+            if times_completed & 1 != 0 {
+                self.direction = !self.direction;
+            }
+        }
         self.clock.reset();
     }
 }
@@ -873,6 +902,27 @@ mod tests {
         last_reported_count: u32,
     }
 
+    /// Utility to create a tween for testing.
+    fn make_test_tween() -> Tween<Transform> {
+        Tween::new(
+            EaseMethod::Linear,
+            Duration::from_secs(1),
+            TransformPositionLens {
+                start: Vec3::ZERO,
+                end: Vec3::ONE,
+            },
+        )
+    }
+
+    /// Utility to create a test environment to tick a tween.
+    fn make_test_env() -> (World, Entity, Transform) {
+        let mut world = World::new();
+        world.insert_resource(Events::<TweenCompleted>::default());
+        let entity = Entity::from_raw(0);
+        let transform = Transform::default();
+        (world, entity, transform)
+    }
+
     #[test]
     fn anim_clock_precision() {
         let duration = Duration::from_millis(1);
@@ -894,7 +944,7 @@ mod tests {
         let mut total_duration = Duration::ZERO;
         for i in 0..10_000_000 {
             let tick = test_ticks[i % test_ticks.len()];
-            times_completed += clock.tick(tick);
+            times_completed += clock.tick(tick).1;
             total_duration += tick;
         }
 
@@ -920,28 +970,26 @@ mod tests {
                 );
 
                 // Create a linear tween over 1 second
-                let mut tween = Tween::new(
-                    EaseMethod::Linear,
-                    Duration::from_secs_f32(1.0),
-                    TransformPositionLens {
-                        start: Vec3::ZERO,
-                        end: Vec3::ONE,
-                    },
-                )
-                .with_direction(*tweening_direction)
-                .with_repeat_count(*count)
-                .with_repeat_strategy(*strategy);
+                let mut tween = make_test_tween()
+                    .with_direction(*tweening_direction)
+                    .with_repeat_count(*count)
+                    .with_repeat_strategy(*strategy);
                 assert_eq!(tween.direction(), *tweening_direction);
                 assert!(tween.on_completed.is_none());
                 assert!(tween.event_data.is_none());
 
-                let dummy_entity = Entity::from_raw(42);
+                let (mut world, entity, mut transform) = make_test_env();
+                let mut event_writer_system_state: SystemState<EventWriter<TweenCompleted>> =
+                    SystemState::new(&mut world);
+                let mut event_reader_system_state: SystemState<EventReader<TweenCompleted>> =
+                    SystemState::new(&mut world);
 
                 // Register callbacks to count started/ended events
                 let callback_monitor = Arc::new(Mutex::new(CallbackMonitor::default()));
                 let cb_mon_ptr = Arc::clone(&callback_monitor);
-                tween.set_completed(move |entity, tween| {
-                    assert_eq!(dummy_entity, entity);
+                let reference_entity = entity;
+                tween.set_completed(move |completed_entity, tween| {
+                    assert_eq!(completed_entity, reference_entity);
                     let mut cb_mon = cb_mon_ptr.lock().unwrap();
                     cb_mon.invoke_count += 1;
                     cb_mon.last_reported_count = tween.times_completed();
@@ -956,16 +1004,7 @@ mod tests {
                 assert!(tween.event_data.is_some());
                 assert_eq!(tween.event_data.unwrap(), USER_DATA);
 
-                // Dummy world and event writer
-                let mut world = World::new();
-                world.insert_resource(Events::<TweenCompleted>::default());
-                let mut event_writer_system_state: SystemState<EventWriter<TweenCompleted>> =
-                    SystemState::new(&mut world);
-                let mut event_reader_system_state: SystemState<EventReader<TweenCompleted>> =
-                    SystemState::new(&mut world);
-
                 // Loop over 2.2 seconds, so greater than one ping-pong loop
-                let mut transform = Transform::default();
                 let tick_duration = Duration::from_secs_f32(0.2);
                 for i in 1..=11 {
                     // Calculate expected values
@@ -989,7 +1028,12 @@ mod tests {
                                 )
                             }
                             RepeatCount::Finite(count) => {
-                                let progress = (i as f32 * 0.2).min(1.0 * *count as f32);
+                                let total_progress = i as f32 * 0.2;
+                                let progress = if total_progress >= *count as f32 {
+                                    1.
+                                } else {
+                                    total_progress.fract()
+                                };
                                 if *strategy == RepeatStrategy::Repeat {
                                     let times_completed = i / 5;
                                     let just_completed = i % 5 == 0;
@@ -1007,8 +1051,8 @@ mod tests {
                                 } else {
                                     let i5 = i % 5;
                                     let times_completed = i / 5;
-                                    let i10 = i % 10;
-                                    let direction = if i10 >= 5 {
+                                    // Once Completed, the direction doesn't change
+                                    let direction = if i >= 5 {
                                         TweeningDirection::Backward
                                     } else {
                                         TweeningDirection::Forward
@@ -1028,7 +1072,7 @@ mod tests {
                                 }
                             }
                             RepeatCount::Infinite => {
-                                let progress = i as f32 * 0.2;
+                                let progress = (i as f32 * 0.2).fract();
                                 if *strategy == RepeatStrategy::Repeat {
                                     let times_completed = i / 5;
                                     let just_completed = i % 5 == 0;
@@ -1079,12 +1123,7 @@ mod tests {
                     // Tick the tween
                     let actual_state = {
                         let mut event_writer = event_writer_system_state.get_mut(&mut world);
-                        tween.tick(
-                            tick_duration,
-                            &mut transform,
-                            dummy_entity,
-                            &mut event_writer,
-                        )
+                        tween.tick(tick_duration, &mut transform, entity, &mut event_writer)
                     };
 
                     // Propagate events
@@ -1112,7 +1151,7 @@ mod tests {
                         if just_completed {
                             assert!(event.is_some());
                             if let Some(event) = event {
-                                assert_eq!(event.entity, dummy_entity);
+                                assert_eq!(event.entity, entity);
                                 assert_eq!(event.user_data, USER_DATA);
                             }
                         } else {
@@ -1157,14 +1196,7 @@ mod tests {
 
     #[test]
     fn tween_dir() {
-        let mut tween = Tween::new(
-            EaseMethod::Linear,
-            Duration::from_secs_f32(1.0),
-            TransformPositionLens {
-                start: Vec3::ZERO,
-                end: Vec3::ONE,
-            },
-        );
+        let mut tween = make_test_tween();
 
         // Default
         assert_eq!(tween.direction(), TweeningDirection::Forward);
@@ -1189,22 +1221,18 @@ mod tests {
         // progress is independent of direction
         assert!(abs_diff_eq(tween.progress(), 0.3, 1e-5));
 
-        // Dummy world and event writer
-        let mut world = World::new();
-        world.insert_resource(Events::<TweenCompleted>::default());
+        let (mut world, entity, mut transform) = make_test_env();
         let mut event_writer_system_state: SystemState<EventWriter<TweenCompleted>> =
             SystemState::new(&mut world);
+        let mut event_writer = event_writer_system_state.get_mut(&mut world);
 
         // Progress always increases alongside the current direction
-        let dummy_entity = Entity::from_raw(0);
-        let mut transform = Transform::default();
-        let mut event_writer = event_writer_system_state.get_mut(&mut world);
         tween.set_direction(TweeningDirection::Backward);
         assert!(abs_diff_eq(tween.progress(), 0.3, 1e-5));
         tween.tick(
             Duration::from_secs_f32(0.1),
             &mut transform,
-            dummy_entity,
+            entity,
             &mut event_writer,
         );
         assert!(abs_diff_eq(tween.progress(), 0.4, 1e-5));
@@ -1231,11 +1259,8 @@ mod tests {
             },
         );
         let mut seq = tween1.then(tween2);
-        let mut transform = Transform::default();
 
-        // Dummy world and event writer
-        let mut world = World::new();
-        world.insert_resource(Events::<TweenCompleted>::default());
+        let (mut world, entity, mut transform) = make_test_env();
         let mut system_state: SystemState<EventWriter<TweenCompleted>> =
             SystemState::new(&mut world);
         let mut event_writer = system_state.get_mut(&mut world);
@@ -1244,7 +1269,7 @@ mod tests {
             let state = seq.tick(
                 Duration::from_secs_f32(0.2),
                 &mut transform,
-                Entity::from_raw(0),
+                entity,
                 &mut event_writer,
             );
             if i < 5 {
@@ -1282,11 +1307,8 @@ mod tests {
             )
             .with_repeat_count(RepeatCount::Finite(1))
         }));
-        let mut transform = Transform::default();
 
-        // Dummy world and event writer
-        let mut world = World::new();
-        world.insert_resource(Events::<TweenCompleted>::default());
+        let (mut world, entity, mut transform) = make_test_env();
         let mut system_state: SystemState<EventWriter<TweenCompleted>> =
             SystemState::new(&mut world);
         let mut event_writer = system_state.get_mut(&mut world);
@@ -1299,7 +1321,7 @@ mod tests {
             seq.tick(
                 Duration::from_secs_f32(delta),
                 &mut transform,
-                Entity::from_raw(0),
+                entity,
                 &mut event_writer,
             );
         }
@@ -1342,7 +1364,7 @@ mod tests {
     fn tracks_tick() {
         let tween1 = Tween::new(
             EaseMethod::Linear,
-            Duration::from_secs_f32(1.),
+            Duration::from_millis(1000),
             TransformPositionLens {
                 start: Vec3::ZERO,
                 end: Vec3::ONE,
@@ -1350,7 +1372,7 @@ mod tests {
         );
         let tween2 = Tween::new(
             EaseMethod::Linear,
-            Duration::from_secs_f32(0.8), // shorter
+            Duration::from_millis(800), // shorter
             TransformRotationLens {
                 start: Quat::IDENTITY,
                 end: Quat::from_rotation_x(90_f32.to_radians()),
@@ -1359,11 +1381,7 @@ mod tests {
         let mut tracks = Tracks::new([tween1, tween2]);
         assert_eq!(tracks.duration(), Duration::from_secs_f32(1.)); // max(1., 0.8)
 
-        let mut transform = Transform::default();
-
-        // Dummy world and event writer
-        let mut world = World::new();
-        world.insert_resource(Events::<TweenCompleted>::default());
+        let (mut world, entity, mut transform) = make_test_env();
         let mut system_state: SystemState<EventWriter<TweenCompleted>> =
             SystemState::new(&mut world);
         let mut event_writer = system_state.get_mut(&mut world);
@@ -1372,7 +1390,7 @@ mod tests {
             let state = tracks.tick(
                 Duration::from_secs_f32(0.2),
                 &mut transform,
-                Entity::from_raw(0),
+                entity,
                 &mut event_writer,
             );
             if i < 5 {
@@ -1448,11 +1466,8 @@ mod tests {
             assert!(tweenable.progress().abs() < 1e-5);
         }
 
-        let mut transform = Transform::default();
-
         // Dummy world and event writer
-        let mut world = World::new();
-        world.insert_resource(Events::<TweenCompleted>::default());
+        let (mut world, entity, mut transform) = make_test_env();
         let mut system_state: SystemState<EventWriter<TweenCompleted>> =
             SystemState::new(&mut world);
         let mut event_writer = system_state.get_mut(&mut world);
@@ -1461,7 +1476,7 @@ mod tests {
             let state = delay.tick(
                 Duration::from_secs_f32(0.2),
                 &mut transform,
-                Entity::from_raw(0),
+                entity,
                 &mut event_writer,
             );
             {
@@ -1482,5 +1497,141 @@ mod tests {
     #[should_panic]
     fn delay_zero_duration_panics() {
         let _ = Delay::new(Duration::ZERO);
+    }
+
+    #[test]
+    fn tween_repeat() {
+        let mut tween = make_test_tween()
+            .with_repeat_count(RepeatCount::Finite(5))
+            .with_repeat_strategy(RepeatStrategy::Repeat);
+
+        assert!(abs_diff_eq(tween.progress(), 0., 1e-5));
+
+        let (mut world, entity, mut transform) = make_test_env();
+        let mut event_writer_system_state: SystemState<EventWriter<TweenCompleted>> =
+            SystemState::new(&mut world);
+        let mut event_writer = event_writer_system_state.get_mut(&mut world);
+
+        // 10%
+        let state = tween.tick(
+            Duration::from_millis(100),
+            &mut transform,
+            entity,
+            &mut event_writer,
+        );
+        assert_eq!(TweenState::Active, state);
+        assert_eq!(0, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 0.1, 1e-5));
+        assert!(transform.translation.abs_diff_eq(Vec3::splat(0.1), 1e-5));
+
+        // 130%
+        let state = tween.tick(
+            Duration::from_millis(1200),
+            &mut transform,
+            entity,
+            &mut event_writer,
+        );
+        assert_eq!(TweenState::Active, state);
+        assert_eq!(1, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 0.3, 1e-5));
+        assert!(transform.translation.abs_diff_eq(Vec3::splat(0.3), 1e-5));
+
+        // 480%
+        let state = tween.tick(
+            Duration::from_millis(3500),
+            &mut transform,
+            entity,
+            &mut event_writer,
+        );
+        assert_eq!(TweenState::Active, state);
+        assert_eq!(4, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 0.8, 1e-5));
+        assert!(transform.translation.abs_diff_eq(Vec3::splat(0.8), 1e-5));
+
+        // 500% - done
+        let state = tween.tick(
+            Duration::from_millis(200),
+            &mut transform,
+            entity,
+            &mut event_writer,
+        );
+        assert_eq!(TweenState::Completed, state);
+        assert_eq!(5, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 1.0, 1e-5));
+        assert!(transform.translation.abs_diff_eq(Vec3::ONE, 1e-5));
+    }
+
+    #[test]
+    fn tween_mirrored_rewind() {
+        let mut tween = make_test_tween()
+            .with_repeat_count(RepeatCount::Finite(4))
+            .with_repeat_strategy(RepeatStrategy::MirroredRepeat);
+
+        assert!(abs_diff_eq(tween.progress(), 0., 1e-5));
+
+        let (mut world, entity, mut transform) = make_test_env();
+        let mut event_writer_system_state: SystemState<EventWriter<TweenCompleted>> =
+            SystemState::new(&mut world);
+        let mut event_writer = event_writer_system_state.get_mut(&mut world);
+
+        // 10%
+        let state = tween.tick(
+            Duration::from_millis(100),
+            &mut transform,
+            entity,
+            &mut event_writer,
+        );
+        assert_eq!(TweenState::Active, state);
+        assert_eq!(TweeningDirection::Forward, tween.direction());
+        assert_eq!(0, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 0.1, 1e-5));
+        assert!(transform.translation.abs_diff_eq(Vec3::splat(0.1), 1e-5));
+
+        // rewind
+        tween.rewind();
+        assert_eq!(TweeningDirection::Forward, tween.direction());
+        assert_eq!(0, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 0., 1e-5));
+        assert!(transform.translation.abs_diff_eq(Vec3::splat(0.1), 1e-5)); // no-op, rewind doesn't apply Lens
+
+        // 120% - mirror
+        let state = tween.tick(
+            Duration::from_millis(1200),
+            &mut transform,
+            entity,
+            &mut event_writer,
+        );
+        assert_eq!(TweeningDirection::Backward, tween.direction());
+        assert_eq!(TweenState::Active, state);
+        assert_eq!(1, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 0.2, 1e-5));
+        assert!(transform.translation.abs_diff_eq(Vec3::splat(0.8), 1e-5));
+
+        // rewind
+        tween.rewind();
+        assert_eq!(TweeningDirection::Forward, tween.direction()); // restored
+        assert_eq!(0, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 0., 1e-5));
+        assert!(transform.translation.abs_diff_eq(Vec3::splat(0.8), 1e-5)); // no-op, rewind doesn't apply Lens
+
+        // 400% - done mirror (because Completed freezes the state)
+        let state = tween.tick(
+            Duration::from_millis(4000),
+            &mut transform,
+            entity,
+            &mut event_writer,
+        );
+        assert_eq!(TweenState::Completed, state);
+        assert_eq!(TweeningDirection::Backward, tween.direction()); // frozen from last loop
+        assert_eq!(4, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 1., 1e-5)); // Completed
+        assert!(transform.translation.abs_diff_eq(Vec3::ZERO, 1e-5));
+
+        // rewind
+        tween.rewind();
+        assert_eq!(TweeningDirection::Forward, tween.direction()); // restored
+        assert_eq!(0, tween.times_completed());
+        assert!(abs_diff_eq(tween.progress(), 0., 1e-5));
+        assert!(transform.translation.abs_diff_eq(Vec3::ZERO, 1e-5)); // no-op, rewind doesn't apply Lens
     }
 }